Finished the parse to read in the XML notes files when an RTIObject is loaded by the program, which the program now does. This is then loaded into the BottomTabPane to display the relevant bookmark sand notes for the currently selected PTMWindow. Created the XML generator, which creates an XML file fo the format of the original RTIViewer, when given an array of Bookmarks. Started implementing this for when new bookmarks/notes are added or deleted for an RTIObject.
Added to stylesheets so that the popup boxes that are generated by the
RTIViewer when an invalid file entry are made are styled with the rest of
the app. Fixed the bug where the position of the light is not at the center
of the specular ball on startup. Made it so light position is maintained
upon resizing the RTIViewer.