HSH files with basis 4 and 9 can now be read. Included Google's guava library for the LittleEndianDataInputStream, used for the HSH parser. Made the default HSH fragment shader work, and created and finished the normals visualisation and specular enhancement fragment shaders.
Decided that making notes backwards compatible with the old RTIViewer is
not going to be possible due to the zoom level on the original viewer
depending on viewer size and monitor resolution, so have decided to just
make bookmarks for this app, so these are now working. Made changing
bookmarks update the UI components, by using ObservableDouble attributes
 for global viewer params, which also nicely eliminated a lot fo ugly
code.